DM-126 Removed unused files in league feature and organized hooks
